Sprint 1 - Please Read!!!
Sprints are supposed to be one week sprints, which means everything
would supposed to be done by Tuesday 11 March. As team lead, I'm
cancelling an official meeting for 11 March, but I do want you to check
in via email or phone call on March 11th to let me know how things are
going.